ref(js): Consolidate dateTime format functions (#51201)
