+++ static/app/components/dateTime.tsx
@@ -2,6 +2,7 @@ import moment from 'moment';
 import momentTimezone from 'moment-timezone';
 
 import ConfigStore from 'sentry/stores/configStore';
+import {getFormat} from 'sentry/utils/dates';
 
 interface Props extends React.HTMLAttributes<HTMLTimeElement> {
   /**
@@ -41,52 +42,6 @@ interface Props extends React.HTMLAttributes<HTMLTimeElement> {
    * For example: "Feb 1" (2022), "Jan 1" (2022), "Dec 31, 2021".
    */
   year?: boolean;
-}
-
-function getDateFormat({year}: Pick<Props, 'year'>) {
-  // "Jan 1, 2022" or "Jan 1"
-  return year ? 'MMM D, YYYY' : 'MMM D';
-}
-
-function getTimeFormat({clock24Hours, seconds, timeZone}) {
-  const substrings = [
-    clock24Hours ? 'HH' : 'h', // hour – "23" (24h format) or "11" (12h format)
-    ':mm', // minute
-    seconds ? ':ss' : '', // second
-    clock24Hours ? '' : ' A', // AM/PM
-    timeZone ? ' z' : '', // time zone
-  ];
-  return substrings.join('');
-}
-
-function getFormat({
-  dateOnly,
-  timeOnly,
-  year,
-  seconds,
-  timeZone,
-  clock24Hours,
-}: Pick<Props, 'dateOnly' | 'timeOnly' | 'year' | 'seconds' | 'timeZone'> & {
-  clock24Hours: boolean;
-}) {
-  if (dateOnly) {
-    return getDateFormat({year});
-  }
-
-  if (timeOnly) {
-    return getTimeFormat({clock24Hours, seconds, timeZone});
-  }
-
-  const dateFormat = getDateFormat({year});
-  const timeFormat = getTimeFormat({
-    clock24Hours,
-    seconds,
-    timeZone,
-  });
-
-  // If the year is shown, then there's already a comma in dateFormat ("Jan 1, 2020"),
-  // so we don't need to add another comma between the date and time
-  return year ? `${dateFormat} ${timeFormat}` : `${dateFormat}, ${timeFormat}`;
 }
 
 function DateTime({
